requirements.txt
PyInstaller==4.9.0
pygame==2.1.2
numpy==1.24.2
pillow==9.4.0